requirements.txt
pyAudioAnalysis==0.3.7
tensorflow==2.5.0
numpy==1.20.1
scikit_learn==0.24.2
torch==1.8.1
matplotlib==3.3.4
pandas==1.2.4
Pillow==8.2.0
split_folders==0.4.3
spectrust==0.1.1
scipy==1.6.2
torchvision==0.9.1
